Tweaks for debian-buster

+++ common/download.sh
@@ -64,10 +64,18 @@ then
   utils.lxc.stop
   echo  >> /var/lib/lxc/${CONTAINER}/config
   echo "# settings for systemd with PID 1:" >> /var/lib/lxc/${CONTAINER}/config
-  echo "lxc.kmsg = 0" >> /var/lib/lxc/${CONTAINER}/config
+  if [ ${DISTRIBUTION} = 'debian' -a ${RELEASE} = 'buster' ];  then
+	echo "Skipping kmsg" >> /var/lib/lxc/${CONTAINER}/config
+  else
+    echo "lxc.kmsg = 0" >> /var/lib/lxc/${CONTAINER}/config
+  fi
   echo "lxc.autodev = 1" >> /var/lib/lxc/${CONTAINER}/config
   utils.lxc.start
-  utils.lxc.attach rm -f /dev/kmsg
+  if [ ${DISTRIBUTION} = 'debian' -a ${RELEASE} = 'buster' ]; then
+	echo "Skipping kmsg again" >> /var/lib/lxc/${CONTAINER}/config
+  else
+    utils.lxc.attach rm -f /dev/kmsg
+  fi
   utils.lxc.stop
 fi
